% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_files_folder}
\alias{check_files_folder}
\title{checks the content of a folder}
\usage{
check_files_folder(path, ext, verbose = TRUE, out = "summary_files")
}
\arguments{
\item{path}{the path}

\item{ext}{a vector of expected extensions of expected files}

\item{verbose}{if TRUE displays some messages related to the check}

\item{out}{\code{all_files} for a list of all files,  \code{summary_files} for a summary or \code{misplaced_files} for a list of misplaced files}
}
\value{
all_files, missplaced files, or  dataframe summary
}
\description{
It checks the extension of files in a folder. It also count the unexpected files.
}
\keyword{internal}
