% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_path}
\alias{check_path}
\title{check if path exists}
\usage{
check_path(path)
}
\arguments{
\item{path}{the path of the folder to check}
}
\value{
the \code{path} if exists, otherwise it throws an Error
}
\description{
Check if the path exists.
}
\keyword{internal}
