% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{choose_direction}
\alias{choose_direction}
\title{choose_direction}
\usage{
choose_direction()
}
\value{
an atomic vector \code{r} or \code{bidirectional}
}
\description{
Use \code{svDialog} to choose is sync left to right or bidirectionally. Use internally
}
