% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftree.R, R/sync.R
\name{clean_up}
\alias{clean_up}
\title{clean_up}
\usage{
clean_up()

clean_up()
}
\description{
Check if folder and file are in the right place documentation is present and if not
cleans it up.

Check if folder and file are in the right place documentation is present and if not
cleans it up.
}
