% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftree.R
\name{create_ftree}
\alias{create_ftree}
\title{create_ftree}
\usage{
create_ftree(path = here::here())
}
\arguments{
\item{path}{where to create the folder (default is \code{here()})}
}
\description{
Create the folder tree and relative documentation files.
}
