% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftree.R
\name{create_overwrite_readme}
\alias{create_overwrite_readme}
\title{create or  overwrite existing readme}
\usage{
create_overwrite_readme(path, type)
}
\arguments{
\item{path}{the path of the parent folder to check}

\item{type}{one of \code{c("data", "ftree")}}
}
\description{
Check if README.txt exists in parent and overwrite it if user agres.
}
