% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftree.R
\name{create_readme}
\alias{create_readme}
\title{create a readme}
\usage{
create_readme(path, type)
}
\arguments{
\item{path}{where to create the folder}

\item{type}{one pf \code{c("data", "f_tree")}}
}
\description{
Create readme file. If \code{path} is \code{data} then the file is created
in \verb{path/data/}.
}
