% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laboR-package.R
\docType{package}
\name{laboR-package}
\alias{laboR}
\alias{laboR-package}
\title{laboR: dades d'atur registrat i contractació de l'Observatori del Treball}
\description{
Descarrega les dades d'atur registrat, demandants d'ocupació i
contractació laboral que publica l'Observatori del Treball i Model
Productiu de la Generalitat de Catalunya.
}
\section{Funcions principals}{

\itemize{
\item Atur registrat: \code{\link[=descarrega_atur_catalunya]{descarrega_atur_catalunya()}},
\code{\link[=descarrega_atur_provincies]{descarrega_atur_provincies()}}, \code{\link[=descarrega_atur_comarques]{descarrega_atur_comarques()}},
\code{\link[=descarrega_atur_municipis]{descarrega_atur_municipis()}}
\item Contractació: \code{\link[=descarrega_contractacio_catalunya]{descarrega_contractacio_catalunya()}},
\code{\link[=descarrega_contractacio_provincies]{descarrega_contractacio_provincies()}},
\code{\link[=descarrega_contractacio_comarques]{descarrega_contractacio_comarques()}},
\code{\link[=descarrega_contractacio_municipis]{descarrega_contractacio_municipis()}}
\item Llistes de referència: \code{\link[=obtenir_municipis]{obtenir_municipis()}}, \code{\link[=obtenir_comarques]{obtenir_comarques()}},
\code{\link[=obtenir_provincies]{obtenir_provincies()}}, \code{\link[=llistar_mesos]{llistar_mesos()}}, \code{\link[=llistar_variables_atur]{llistar_variables_atur()}},
\code{\link[=llistar_variables_contractacio]{llistar_variables_contractacio()}}
}
}

\section{Secret estadístic}{

En municipis de menys de 20.000 habitants, les cel·les amb valors petits
(menys de 4) se suprimeixen per secret estadístic i el servidor les marca
amb "#". El paquet les converteix sistemàticament en \code{NA}, de manera que
les columnes de valors són sempre numèriques.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gerardreverte/laboR}
  \item Report bugs at \url{https://github.com/gerardreverte/laboR/issues}
}

}
\author{
\strong{Maintainer}: Gerard Reverté Calvet \email{gerardrc@gmail.com}

}
\keyword{internal}
