% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{labor_sync}
\alias{labor_sync}
\title{labor sync}
\usage{
labor_sync(bidirectional = FALSE, exclude_files = "default")
}
\arguments{
\item{bidirectional}{boolean}

\item{exclude_files}{list of file to exclude. if "default" list all hidden files in parent, and folders in \code{renv}.}
}
\description{
sync the project folder with a local folder. The origin is alwaysy the parental director
}
\details{
use \code{rsync} to sync and \code{here} to identify parental directory
}
