% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comu.R
\name{obtenir_provincies}
\alias{obtenir_provincies}
\title{Llista les províncies catalanes disponibles}
\usage{
obtenir_provincies(verbose = FALSE)
}
\arguments{
\item{verbose}{Mostra missatges de progrés (per defecte \code{FALSE}).}
}
\value{
Un \code{data.frame} amb les columnes \code{codi}, \code{nom}, \code{id_intern} i
\code{value_str} per a les 4 províncies catalanes, o \code{NULL} si el servidor
no respon.
}
\description{
Llista les províncies catalanes disponibles
}
\examples{
\dontrun{
provincies <- obtenir_provincies()
}
}
