% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_folders.R
\name{path_check}
\alias{path_check}
\title{check if path exists}
\usage{
path_check(path)
}
\arguments{
\item{path}{the path of the folder to check}
}
\value{
the \code{path} if exists, otherwise it throws an Error
}
\description{
Check if the path exists.
}
