% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{purlify}
\alias{purlify}
\title{purlify}
\usage{
purlify(file)
}
\arguments{
\item{file}{path to file to \code{purlify}}
}
\value{
\code{R} script  with layout common in the lab. The text is commented.
}
\description{
Transform an \verb{RMarkdown notebook} in a \verb{R script}
}
\details{
the function relies on \code{knitr::purl} for the muscle work, and then clean the file up further.
}
