% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_DS_Store}
\alias{remove_DS_Store}
\title{Remove DS Stores}
\usage{
remove_DS_Store()
}
\description{
Remove \code{.DS_Store} files and modify .git_ignore to avoid tracking of those files. Based on
\href{https://stackoverflow.com/questions/107701/how-can-i-remove-ds-store-files-from-a-git-repository}{this post.}:
}
