% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_file}
\alias{remove_file}
\title{Remove File}
\usage{
remove_file(file_name)
}
\description{
Function that remove a file recursivelly and added it to .gitignore to avoid tracking of those files. Based on
\href{https://stackoverflow.com/questions/107701/how-can-i-remove-ds-store-files-from-a-git-repository}{this post.}:
}
