% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftree.R
\name{remove_ftree}
\alias{remove_ftree}
\title{remove_ftree}
\usage{
remove_ftree(path = here::here())
}
\arguments{
\item{path}{where to create the folder (defaut is \code{here()})}
}
\description{
remove the folder tree and relative documentation files
}
