% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{remove_lfolder}
\alias{remove_lfolder}
\title{remove_lfolder}
\usage{
remove_lfolder(name_lf, path = NULL, inter = TRUE)
}
\arguments{
\item{name_lf}{name of lab folder}

\item{path}{path to folder}

\item{inter}{boolean for using \code{svDialogs} to inteactively select folder}
}
\value{

}
\description{
Remove lab folder and relative documentation in the appropriate file
}
