% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrive_copy_files_pkg}
\alias{retrive_copy_files_pkg}
\title{Retrieve file and copy it}
\usage{
retrive_copy_files_pkg(file_names, folder_out)
}
\arguments{
\item{file_names}{Name of the files.}

\item{folder_out}{Where to copy it.}
}
\description{
Find file in {labor} package and copy it.
}
