% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{set_sync_lab}
\alias{set_sync_lab}
\title{Set the destination of the syncing}
\usage{
set_sync_lab()
}
\description{
Interactively set the folder to sync. That info will be recorder in \code{.labor_destination} in the parental directory
}
