% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_folders.R
\name{setup_lab_project}
\alias{setup_lab_project}
\title{labtree wizard.
#'
#' Choose folders to use in the labtree.}
\usage{
setup_lab_project(
  use_targets = FALSE,
  path_project = here::here(),
  pkg_to_install = c("devtools", "here", "reticulate", "tidyverse", "usethis"),
  files_git_rm = c(".DS_Store", "._.DS_Store", "._.*"),
  proj_name,
  use_python = TRUE
)
}
\arguments{
\item{use_targets}{If FALSE, do not use targets. If "local" install targets packages, add \verb{_target.R} and \code{function.R} files
, and update \code{README.Rmd}. If "cluster" add other config files for use with \code{slurm} and update \code{README.Rmd} accordingly.}

\item{pkg_to_install}{A vector of packages to install.}

\item{proj_name}{If \code{use_targets = "cluster"}, name of the project. Used to rename file.Rprj}

\item{use_python}{If TRUE launch \code{renv::use_python()}}
}
\description{

}
\section{do you want to use the defaul labtree?}{

}

