% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{sync_lab}
\alias{sync_lab}
\title{labor sync}
\usage{
sync_lab(
  direction = "or_de",
  exclude_files = "default",
  rsync_flags = "-avtuP"
)
}
\arguments{
\item{direction}{one of \verb{c("or_de", "de_or", "bidir") meaning }origin to destination\verb{, }destination to origin\code{or}bidirectional\verb{. Defalut is }or_de`.}

\item{exclude_files}{list of files to exclude. If "default", it will  not sync hidden files in parent as well as  folders in \code{renv}. If
"none" will sync everything.}

\item{rsync_flags}{flag to use wit \code{rsync}.  Default is \code{-avtuP}. Check \href{https://ss64.com/bash/rsync_options.html}{rsync page}
for the complete list of options.}
}
\description{
sync the project folder with the local folder set by \code{set_lab_sync}. The origin is always the parental director found
by \code{here::here()}.
}
\details{
use \code{rsync} to sync and \code{here} to identify parental directory
}
