% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{sync_proj}
\alias{sync_proj}
\title{sync_proj}
\usage{
sync_proj(direction, path)
}
\arguments{
\item{direction}{one of \code{c("l", "r", "bd")}}
}
\description{
sync the project folder with a local folder
}
\details{
use rsync
}
