% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{sync_proj_loc}
\alias{sync_proj_loc}
\title{sync_proj_local}
\usage{
sync_proj_loc(
  direction = "r",
  inter = "FALSE",
  origin = "man/",
  destination = "~/Desktop/man"
)
}
\arguments{
\item{direction}{one of \code{c("l_to_r", "bd")}}

\item{inter}{boolean interactive choosing}

\item{origin}{path to origin folder (default is wd)}

\item{destination}{path to destination}
}
\description{
sync the project folder with a local folder
}
\details{
use rsync
}
