% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{who_wants_to_talk}
\alias{who_wants_to_talk}
\title{who_wants_to_talk}
\usage{
who_wants_to_talk(lab = "marchionni")
}
\arguments{
\item{lab}{one of \verb{marchionni", "wheelan", "joint}}
}
\value{
a message
}
\description{
who_wants_to_talk
}
