% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_folders.R
\name{create_labtree}
\alias{create_labtree}
\title{Create labtree.}
\usage{
create_labtree(path_project)
}
\arguments{
\item{path}{Where to create the folder (default is \code{here()}).}
}
\description{
Create the folder tree and relative documentation files. If folders are already present, it will ask
which one to overwrite.
}
\keyword{internal}
