% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{labor_sync}
\alias{labor_sync}
\title{labor sync}
\usage{
labor_sync(
  direction = "or_de",
  exclude_files = "default",
  rsync_flags = "-avtuP"
)
}
\arguments{
\item{direction}{one of \verb{c("or_de", "de_or", "bidir") that stands origin to destination, destination to origin or bidirectional. Defalut is }or_de`}

\item{exclude_files}{list of file to exclude. If "default" do not sync hidden files in parent, and folders in \code{renv}. If
"none" will sync everything.}

\item{rsync_flags}{flag to use wit \code{rsync}, default is \code{-avtuP}. Check \href{https://ss64.com/bash/rsync_options.html}{rsync page}
for the complete list of options.}
}
\description{
sync the project folder with a local folder. The origin is alwaysy the parental director
}
\details{
use \code{rsync} to sync and \code{here} to identify parental directory
}
