% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_folders.R
\name{remove_labtree}
\alias{remove_labtree}
\title{Remove labtree.}
\usage{
remove_labtree(project_oath = here::here())
}
\arguments{
\item{path}{where to create the folder (defaut is \code{here()})}
}
\description{
Femove the folder tree and relative documentation files.
}
