% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_folders.R
\name{setup_lab_project}
\alias{setup_lab_project}
\title{set up project}
\usage{
setup_lab_project(use_targets = FALSE)
}
\arguments{
\item{use_targets}{If TRUE creates target scripts and}
}
\description{
Install some packages commonly used and create the project structure.
}
