% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_folders.R
\name{setup_lab_project}
\alias{setup_lab_project}
\title{set up project}
\usage{
setup_lab_project(
  path_project,
  use_targets = FALSE,
  pkg_to_install = c("BiocManager", "devtools", "here", "lintr", "languageserver",
    "renv", "targets", "tidyverse", "usethis"),
  use_git = TRUE
)
}
\arguments{
\item{path_project}{Where to setup the project.}

\item{use_targets}{If FALSE, do not use targets. If "local" install targets packages, add \verb{_target.R} and \code{function.R} files
, and update \code{README.Rmd}. If "cluster" add other config files for use with \code{slurm} and update \code{README.Rmd} accordingly.}

\item{pkg_to_install}{A vector of packages to install.}

\item{use_git}{If TRUE install git}

\item{files_git_rm}{A vector of files name to remove.}
}
\description{
Install some packages commonly used, initiate \code{renv} and create the project structure.
}
