% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{sync_project_drop}
\alias{sync_project_drop}
\title{sync project drop}
\usage{
sync_project_drop(direction = "local_to_drop", inter = "FALSE", local, dropbox)
}
\arguments{
\item{direction}{one of `c("local_to_drop", "bd", "drop_to_local")}

\item{inter}{if interaction is true choose the direction interactivelly}

\item{local}{path to origin folder (default is wd)}

\item{dropbox}{path to destination}
}
\description{
sync the project folder with dropbox folder
}
\details{
use rdrop
}
